Creada estructura nueva para home y  USER, y las rutas tipo /api
