+++ app/Http/Controllers/HomeController.php
@@ -23,6 +23,7 @@ class HomeController extends Controller
      */
     public function index()
     {
-        return view('home');
+        // return view('home');
+        return view('welcome');
     }
 }
