+++ resources/js/components/Mupp.vue
@@ -1,0 +1,1 @@
+<!-- raiz de la app de vue -->
