+++ webpack.mix.js
@@ -31,6 +31,8 @@ var webpackConfig ={
 mix.webpackConfig(webpackConfig);
 
 mix.js('resources/js/app.js', 'public/js')
-    .vue()
-    .sass('resources/sass/app.scss', 'public/css')
+
+    .sass('resources/sass/app.scss', 'public/css', [
+
+    ]).vue()
